+++ CHANGELOG.md
@@ -24,7 +24,8 @@
 - Key definitions
   - [BREAKING CHANGE] Switch launcher key from config `"bract.core.launcher"` to context `:bract.core/launcher`
   - Change launcher parser from `str->var` to `any->fn`
-  - [BREAKING CHANGE] Remove exports config keydef (use variable instead)
+  - [BREAKING CHANGE] Remove exports config keydef (in favor of using variable instead)
+  - [BREAKING CHANGE] Set default context file to `bract-context.edn`
 - Inducers
   - [BREAKING CHANGE] Have `export-as-sysprops` and `unexport-sysprops` accept export-key/value arguments
 - [Todo - BREAKING CHANGE] Drop bract.core.type and bract.core.impl in favour of a multimethod (for flexibility)