+++ CHANGELOG.md
@@ -28,8 +28,10 @@
   - Set default value for config key `config-inducers` to `[]`
   - Change launcher parser from `str->var` to `any->fn`
 - Inducers
-  - [BREAKING CHANGE] Have `export-as-sysprops` and `unexport-sysprops` accept export-key/value arguments
-  - Fix issue where vector and map arguments are misconstrued as functions upon parsing
+  - [BREAKING CHANGE] Have inducers dealing with system properties accept export-key/value arguments
+    - `export-as-sysprops`
+    - `unexport-sysprops`
+  - Fix issue where vector and map arguments are misinterpreted as functions upon parsing
 - CLI entrypoint
   - Add `bract.core.main` namespace for CLI entry point
     - Define root inducers